refactor tests for orthogonality
this ended up reducing the number of tests

the rationale is: I want to avoid a combinatorial explosion of test
cases so instead of testing

- if user name is in sudoers user list and correct password = OK
- if user name is in sudoers user list and incorrect password = FAIL
- if user name is in sudoers user list and no password = FAIL
- if users group name is in sudoers user list and correct password is provided
- if users group name is in sudoers user list and incorrect password is provided
- if users group name is in sudoers user list and no password is provided
(.. ~16 tests  ..)

I test

GIVEN valid password auth (or NOPASSWD, it doesn't matter much which)
- no match in sudoers user list = FAIL
- ALL in sudoers user list = OK
- if user name is in sudoers user list = OK
- if users group name is in sudoers user list = OK
- if user id is in sudoers user list = OK
- if users group id is in sudoers user list = OK

and separately I test

GIVEN match in sudoers match list
- if correct password = OK
- if incorrect password = FAIL
- if no password = FAIL

(9 tests total)

note that because these are end to end tests we are always testing the
whole pipeline. it's just that because we have knowledge about the
order of the components of the pipeline (e.g. sudoers gets checked
before the password prompt) we can have orthogonal test cases that
assume the previous / next component will succeed and still have good
coverage.