Use `chdir` in `pre-exec`
These changes set the working directory of the command to be executed by
sudo during the `pre-exec` stage instead of calling
`Command::current_dir` and print an user error if this operation fails.

This means that we no longer check if the target directory exists before
chaging the working directory.