fix potential deadlock in pre-exec
pre-exec runs after fork, so allocations are forbidden. Move the
allocation outside of the closure. The closure itself is leaked though,
so it's fine to close over c_path --- it will be dropped in parent, and
will be leaked in the child

Practically, I don't think anything _particularly_ bad can happen, but
deadlock in multithreaded programs is a very real possibility!

What could happen is:

* thread A takes a global mutex guarding the allocator on the slow path
* thread B forks out an address space with the mutex in the locked state
* the forked process allocates (in CString::new) and gets stuck.
* thread B gets stuck waiting on the pipe to this new process

Another problem here is logging --- it could also allocate, and uses the
file descriptor from the child process. Ideally, we'd log in the parent
process after exec, but we can't unambiguously figure out from the exit
status that chdir is the problem.