refactor: add SudoString and SudoPath
these are versions of libstd's String and PathBuf that contain no null
bytes and thus can be easily converted into CStr

SudoPath has the additional invariant that it is UTF-8 encoded. SudoPath
is used to represent CWD which appears in the sudoers file. sudo-rs
only accepts UTF-8 encoded sudoers files

closes #748