Ensure the si_pid field is only accessed when initialized
The si_pid field would be logged even when it is uninitialized. It also
obvious that it isn't possible to trick sudo into thinking that the
signal it received was from itself due to the uninitialized value
it's own pid by chance. This is not possible as the check is prefixed by
is_user_signaled and all signals where is_user_signaled are true have
si_pid initialized. Using an option ensures that this check is never
forgotten.