Remove broken test
This removes a test that seemed like it checks that sudo doesn't set PS1
even when SUDO_PS1 is set in the case of a login shell. As it turns out
however, this behavior is caused by bash not inheriting PS1 to it's
children. If another shell is used, PS1 will be shown to be passed by
sudo to the elevated process even with og sudo.