Add a testing only feature for running as unprivileged user
This feature must never be used in production. To ensure this, it needs
to be enabled using RUSTFLAGS rather than as cargo feature, there is a
compile error for using it with debug assertions enabled and we also
check for setuid and being root at runtime and refuse to run in either
case.