Fuse the three loops in converse
This simplifies the code and makes it easier to directly return the
response from Converser::handle_message in the future.

This will leak some memory when there are multiple messages in a single
conversation and handling one of the messages returns an error. This
shouldn't matter for sudo-rs however as we would exit shortly after the
error anyway.