Add basic FreeBSD CI using Cirrus CI (#923)
This currently only runs unit tests and not the compliance tests.

Cirrus CI is the only CI provider I'm aware of which supports FreeBSD
out of the box. It does have the disadvantage though that on the free
tier it can take several minutes before a CI job actually starts.

---------

Co-authored-by: Marc R. Schoolderman <info@squell.net>