Use docker run --rm + docker kill instead of docker rm
Unlike what the comment on docker rm suggested, previously we would
actually block on docker rm completing. Using docker run --rm +
docker kill only blocks on killing the container while running the slow
cleanup in the background. This almost halfs the time it takes to run
the compliance tests with podman on Linux. In addition if you kill the
compliance test executable itself, the containers will now be gone
entirely as soon as you kill the containers rather than leaving data on
the disk.