Rewrite `.fold()` into something that is immediately readable (#1052)
Based on #1036, rewriting the `.fold()`.

Commits are done step-by-step to show that the rewrite behaves
equivalently (except for 75554ac, but that is fixed in the final
beautiful result).

You can review this commit by commit (this will show you my "I'm not
trying to be clever" careful editing of the code), but you can also just
look at the final result which will be very easy to review.

Obviously we're going to squash these commits.

A hearty thanks to @cikzh for rubber ducking.