Revert "Implement support for NOEXEC (#1073)"
This is spawning a thread before fork, which is UB. This will need to be
fixed, but given that we plan to do a release in a couple of days, let's
remove NOEXEC support for now and re-add fixed NOEXEC support after the
release.

This reverts commit ca76a705e72dab31e650704be8ccae14f971d793, reversing
changes made to 5ca07b1b8f8e157aa58c2d825118e4a4f45c2e0d.