Fix a panic when exec fails with use_pty enabled
Previously this would panic as the parent process expected the monitor
process to have alreqady broken out of the event loop before sending
back an error message. Which then resulted in an edge monitor message,
which is only expected to be received outside the event loop, causing a
panic. By using set_break we still cause the error to be sent to the
parent process, while ensuring that we are no longer inside the event
loop of the monitor when doing so.