Avoid interleaving logs from different processes
Previously there would be multiple write calls. One for the prefix and
one or more for the log arguments. Between each write call there can be
another process logging part of a log message. By instead buffering the
entire log and writing it as a single write we avoid this interleaving.