Use auth user rather than current user in timestamp records
This fixes an auth bypass when combining sudo -u with Defaults targetpw.
Previously we would reuse the timestamp records between different target
users even when we are asking for the password of the target user, so
knowing the password of one user would allow creating a record which is
valid for logging in as another user.

This doesn't require bumping the timestamp file format version as the
record layout remains identical other than swapping the current user's
uid with the auth user's uid. This means that existing timestamp records
produced by a buggy sudo-rs only allow skipping the authentication check
while they are valid when the password of the current user is necessary.
When targetpw is enabled this would only be the case when you try to run
sudo -u $(id -u), which doesn't gain you any new privileges.