Don't open /dev/tty twice
There is no need to do so and this refactor clarifies exactly when
HiddenInput::new() returns Some. With this we no longer disable feedback
on /dev/tty when we are reading a password from a pipe and we disable
feedback on the correct tty if stdin is a different tty than /dev/tty.
Once we implement askpass this also prevents interfering with the
askpass program if it reads from the tty.