Fix a lint warning introduced on nightly
The new function_casts_as_integer lint warns on direct casts from
function items to integers as those sometimes indicate a bug in the
program. In our case the cast is however correct, so silence the lint by
first casting to *const () as suggested by the lint.

Also fix a todo now that we use a newer clippy version.