Don't set the MAIL env var
On Linux pretty much every program that used to read it no longer exists
in distros as far as we can tell. And the few that still do, already
support getting the default value that sudo-rs would have set anyway. In
addition neither a regular gui session nor openssh sets this env var.

We will still respect setting the env var if it is in env_keep,
env_check or set by PAM through for example pam_mail like on the default
PAM config of Debian for su and login.