Remove rust caching from most CI jobs
sudo-rs only has 2 dependencies remaining. Together they should compile
about as fast as trying to restore a cache, so there is no point in
keeping those caches.

This still keeps the rust caching for the compliance tests. For them it
probably doesn't save a lot of time, but these are the slowest tests, so
every little bit helps. And unlike sudo-rs itself, there are several
deps.