Remove a few unwrap()'s (#1383)
Inspired by some recent unasked-for advice aimed at a certain company, I
decided to have a look at all the nice presents elves put in our code
base for us to unwrap. Most of them are nice and proper (e.g. the line
on line 379 in noexec.rs stands out as a perfect example).

But I found a few in which we might as well change the code (although it
will have no noticable effect).

According to experts, this makes our code less dangerous.